chore: bump

+++ www-client/vivaldi/vivaldi-7.1.3570.42.ebuild
@@ -1,4 +1,4 @@
-# Copyright 1999-2025 Gentoo Authors
+# Copyright 1999-2024 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
 EAPI=8
@@ -109,9 +109,8 @@ S="${WORKDIR}"
 LICENSE="Vivaldi"
 SLOT="0"
 KEYWORDS="-* ~amd64 ~arm ~arm64"
-IUSE="ffmpeg-chromium gtk proprietary-codecs qt6 widevine"
+IUSE="gtk proprietary-codecs qt5 qt6 widevine"
 RESTRICT="bindist mirror"
-#REQUIRED_USE="ffmpeg-chromium? ( proprietary-codecs )"
 
 RDEPEND="
 	>=app-accessibility/at-spi2-core-2.46.0:2
@@ -133,11 +132,15 @@ RDEPEND="
 	x11-libs/libXfixes
 	x11-libs/libxkbcommon
 	x11-libs/libXrandr
-	x11-libs/pango
+	x11-libs/pango[X]
 	gtk? ( gui-libs/gtk:4 x11-libs/gtk+:3 )
 	proprietary-codecs? (
-		!ffmpeg-chromium? ( >=media-video/ffmpeg-6.1-r1:0/58.60.60[chromium] )
-		ffmpeg-chromium? ( media-video/ffmpeg-chromium:${CHROMIUM_VERSION} )
+		>=media-video/ffmpeg-7.0.2:0/58.60.60[chromium]
+	)
+	qt5? (
+		dev-qt/qtcore:5
+		dev-qt/qtgui:5
+		dev-qt/qtwidgets:5
 	)
 	qt6? ( dev-qt/qtbase:6[gui,widgets] )
 	widevine? ( www-plugins/chrome-binary-plugins )
@@ -179,8 +182,9 @@ src_prepare() {
 		rmdir ${VIVALDI_HOME}/lib || die
 	fi
 
-	# Qt5 is obsolete now.
-	rm ${VIVALDI_HOME}/libqt5_shim.so || die
+	if ! use qt5; then
+		rm ${VIVALDI_HOME}/libqt5_shim.so || die
+	fi
 
 	if ! use qt6; then
 		rm ${VIVALDI_HOME}/libqt6_shim.so || die
@@ -202,7 +206,7 @@ src_install() {
 	done
 
 	if use proprietary-codecs; then
-		dosym ../../usr/$(get_libdir)/chromium/libffmpeg.so$(usex ffmpeg-chromium .${CHROMIUM_VERSION} "") \
+		dosym ../../usr/$(get_libdir)/chromium/libffmpeg.so \
 			/${VIVALDI_HOME}/libffmpeg.so.$(ver_cut 1-2)
 	fi
